+++ src/clustering/administration/main/command_line.cc
@@ -74,7 +74,7 @@ void run_rethinkdb_create(const std::string &filepath, const std::string &machin
 
 std::set<peer_address_t> look_up_peers_addresses(std::vector<host_and_port_t> names) {
     std::set<peer_address_t> peers;
-    for (int i = 0; i < (int)names.size(); ++i) {
+    for (size_t i = 0; i < names.size(); ++i) {
         peers.insert(peer_address_t(ip_address_t(names[i].host), names[i].port));
     }
     return peers;
